@@ -1,7 +1,7 @@
 {
-  "name": "mean3",
+  "name": "means",
   "version": "1.0.0",
-  "description": "MongoDB + Angular 5 + ExpressJS on TypeScript + NodeJS server",
+  "description": "MongoDB + Angular 6 + ExpressJS on TypeScript + NodeJS server",
   "main": "app/app.js",
   "scripts": {
     "backend": "gulp",
@@ -17,25 +17,26 @@
   "author": "Jaroslaw Frydrych <jarek.frydrych@gmail.com> (github.com/jar5551)",
   "license": "ISC",
   "dependencies": {
-    "@angular/animations": "^5.2.0",
-    "@angular/common": "^5.2.0",
-    "@angular/compiler": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/forms": "^5.2.0",
-    "@angular/http": "^5.2.0",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@angular/router": "^5.2.0",
+    "@angular/animations": "6.0.0",
+    "@angular/common": "6.0.0",
+    "@angular/compiler": "6.0.0",
+    "@angular/core": "6.0.0",
+    "@angular/forms": "6.0.0",
+    "@angular/http": "6.0.0",
+    "@angular/platform-browser": "6.0.0",
+    "@angular/platform-browser-dynamic": "6.0.0",
+    "@angular/router": "6.0.0",
     "core-js": "^2.4.1",
     "express-jwt": "^5.3.1",
     "normalize.css": "^8.0.0",
-    "rxjs": "^5.5.6",
-    "zone.js": "^0.8.19"
+    "rxjs": "^6.1.0",
+    "zone.js": "~0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "~1.7.4",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/language-service": "^5.2.0",
+    "@angular-devkit/build-angular": "~0.6.0",
+    "@angular/cli": "6.0.0",
+    "@angular/compiler-cli": "6.0.0",
+    "@angular/language-service": "6.0.0",
     "@types/body-parser": "^1.16.8",
     "@types/express": "^4.11.1",
     "@types/jasmine": "~2.8.3",
@@ -48,7 +49,7 @@
     "babel-preset-react": "^6.16.0",
     "babel-preset-stage-0": "^6.16.0",
     "babel-register": "^6.26.0",
-    "bcrypt-nodejs": "0.0.3",
+    "bcrypt": "^2.0.1",
     "body-parser": "^1.18.2",
     "codelyzer": "^4.0.1",
     "cookie-parser": "^1.4.3",
@@ -79,7 +80,7 @@
     "ts-node": "~4.1.0",
     "tslint": "^5.9.1",
     "typegoose": "^5.0.0",
-    "typescript": "~2.5.3",
+    "typescript": "2.7.2",
     "typings": "^2.1.1"
   }
 }
