@@ -1,7 +1,8 @@
 import {Injectable} from '@angular/core';
 import {CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot, Router} from '@angular/router';
-import {Observable} from 'rxjs/Observable';
+import {Observable} from 'rxjs';
 import {UserService} from './user/user.service';
+import {map} from 'rxjs/operators';
 
 @Injectable()
 export class AuthGuard implements CanActivate {
@@ -13,11 +14,11 @@ export class AuthGuard implements CanActivate {
     canActivate(route: ActivatedRouteSnapshot,
                 state: RouterStateSnapshot): Observable<boolean> {
         return this.userService.checkCurrentUser()
-            .map((value: boolean) => {
+            .pipe(map((value: boolean) => {
                 if (!value) {
                     this.router.navigate(['/']);
                 }
                 return value;
-            });
+            }));
     }
 }
