@@ -1,8 +1,9 @@
 import {Injectable} from '@angular/core';
 import {CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot, Router} from '@angular/router';
-import {Observable} from 'rxjs/Observable';
+import {Observable} from 'rxjs';
 import {UserService} from './user/user.service';
 import {CoreService} from '../core/core.service';
+import {map} from 'rxjs/operators';
 
 @Injectable()
 export class LoginGuard implements CanActivate {
@@ -15,11 +16,11 @@ export class LoginGuard implements CanActivate {
     canActivate(route: ActivatedRouteSnapshot,
                 state: RouterStateSnapshot): Observable<boolean> {
         return this.userService.checkCurrentUser()
-            .map((value: boolean) => {
+            .pipe(map((value: boolean) => {
                 if (value) {
                     this.router.navigate([this.coreService.homePage]);
                 }
                 return !value;
-            });
+            }));
     }
 }
