@@ -1,6 +1,4 @@
 "use strict";
-function __export(m) {
-    for (var p in m) if (!exports.hasOwnProperty(p)) exports[p] = m[p];
-}
 Object.defineProperty(exports, "__esModule", { value: true });
-__export(require("./core/core.controller"));
+var core_controller_1 = require("./core/core.controller");
+exports.asdHandler = core_controller_1.asdHandler;
