@@ -1,1 +1,1 @@
-export * from './security/security.controller';
+export {loginHandler, logoutHandler, userNewHandler, currentUserQuery} from './security/security.controller';
